requirements.txt
pygame>=2.6.0
numpy>=1.24.0
Pillow>=10.0.0
neat-python>=0.92
pandas>=2.0.0
matplotlib>=3.7.0